First crack at implementing the logging.
There will be a 'console' log which will tee writes to local stdout, a
console log file, and via the GearmanJob sendData call to forward the
console output to the client.

It is debateable whether or not we will use this log for non-build
related stuff.  It may be more prudent to have an application log using
a standard logging framework.